refactor: classes moved to SmkPsiElements.kt file
