+++ src/main/kotlin/com/jetbrains/snakecharm/lang/psi/SmkPsiElements.kt
@@ -7,6 +7,7 @@ import com.intellij.psi.StubBasedPsiElement
 import com.jetbrains.python.PyTokenTypes
 import com.jetbrains.python.codeInsight.controlflow.ScopeOwner
 import com.jetbrains.python.psi.*
+import com.jetbrains.python.psi.impl.PyElementImpl
 import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.WILDCARDS_DEFINING_SECTIONS_KEYWORDS
 import com.jetbrains.snakecharm.codeInsight.SnakemakeAPI.WILDCARDS_EXPANDING_SECTIONS_KEYWORDS
 import com.jetbrains.snakecharm.lang.psi.stubs.*
@@ -56,6 +57,9 @@ interface SmkUse : SmkRuleOrCheckpoint, StubBasedPsiElement<SmkUseStub> {
      */
     fun nameIdentifierIsWildcard(): Boolean
 }
+
+class SmkUseNameIdentifier(node: ASTNode) : PyElementImpl(node)
+class SmkImportedRulesNames(node: ASTNode) : PyElementImpl(node)
 
 interface SmkRuleOrCheckpointArgsSection : SmkArgsSection, PyTypedElement { // PyNamedElementContainer
     /**
