--- src/main/kotlin/com/jetbrains/snakecharm/lang/psi/impl/SmkUseNameIdentifier.kt
@@ -1,8 +1,0 @@
-package com.jetbrains.snakecharm.lang.psi.impl
-
-import com.intellij.lang.ASTNode
-import com.jetbrains.python.psi.impl.PyElementImpl
-
-class SmkUseNameIdentifier(node: ASTNode) : PyElementImpl(node)
-
-class SmkImportedRulesNames(node: ASTNode) : PyElementImpl(node)
